WDL_Resampler: when in ideal-sinc mode, avoid allocating/using extra duplicate slice
The last (duplicate) slice is a reversed (or: offset by one sample, same
thing) copy of the first slice, which for purposes of ideal-sinc, was
incorrect.

It is preserved for sinc-interpolated mode. We could also add a sample
before the first slice, and use that instead of the last slice, if we
wanted to save a tiny bit of memory in exchange for some extra logic. -- from 89b28942